+++ src/components/modules/FormAddNotice/helpers/AddIcon.jsx
