+++ src/components/modules/FormAddNotice/helpers/getCategory.jsx
