+++ src/components/modules/FormAddNotice/helpers/getWidth.jsx
