+++ src/components/modules/FormAddNotice/index.js
@@ -1,0 +1,3 @@
+export * from './FormAddNotice';
+export * from './FormStepOne';
+export * from './FormStepTwo';
