+++ src/page/AddNoticePage/index.js
@@ -1,0 +1,1 @@
+export { default as AddNoticePage } from './AddNoticePage';
